refactor: multiple parser modules
